backend/requirements/base.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.3.0
cloudinary==1.40.0
django-cloudinary-storage==0.3.0
whitenoise==6.6.0
